+++ pypowerwall/tedapi/ted_api_messages/device_controller_message.py
@@ -1,9 +1,8 @@
 import logging
 import json
-import os
 from .. import tedapi_pb2
 from ..exceptions import PyPowerwallTEDAPIException
-from .TEDAPIMessage import TEDAPIMessage
+from .tedapi_message import TEDAPIMessage
 
 log = logging.getLogger(__name__)
