+++ pypowerwall/tedapi/ted_api_messages/firmware_message.py
@@ -1,6 +1,6 @@
 import logging
 from .. import tedapi_pb2
-from .TEDAPIMessage import TEDAPIMessage
+from .tedapi_message import TEDAPIMessage
 
 log = logging.getLogger(__name__)
