+++ pypowerwall/tedapi/ted_api_messages/gateway_status_message.py
@@ -2,7 +2,7 @@ import logging
 import json
 from .. import tedapi_pb2
 from ..exceptions import PyPowerwallTEDAPIException
-from .TEDAPIMessage import TEDAPIMessage
+from .tedapi_message import TEDAPIMessage
 
 log = logging.getLogger(__name__)
