+++ pypowerwall/tedapi/ted_api_messages/tedapi_message.py
